Refactor complete. changed a few image names and improved font of download button
