+++ personal-portfolio/src/App.jsx
@@ -1,6 +1,7 @@
 import Home from './Pages/Landing Page/Home'
 import WorkFlow from './Pages/Workflow Page/WorkFlow'
 import Contact from './Pages/Contact Page/Contact'
+import Me from './assets/Me.png'
 import Gallery from './Pages/Gallery/Gallery'
 import './App.css'
 import React from 'react'
@@ -14,7 +15,7 @@ function App() {
     title:"Hello, I'm Manqoba Nathan Mtetwa",
     desr: "I am an upcoming Designer and Developer based in Johannesburg. I zealously strive towards carefully designing beautiful UIs, writing clean and efficient code for web-applications and managing technical projects. As a fresh graduate from Belgium Campus, I am eager to land my first role as a",
     p:"Junior Developer.",
-    CVlink: "-------", img: './src/assets/Me.png'
+    CVlink: "-------", img:Me,
   }
   return (
     <>
